+++ src/main.rs
@@ -9,9 +9,12 @@ mod utils {
 }
 
 fn main() {
-    let matches = App::new("jokeyrhyme-dotfiles")
+    let matches = App::new(env!("CARGO_PKG_NAME"))
         .version(env!("CARGO_PKG_VERSION"))
-        .subcommand(SubCommand::with_name("sync"))
+        .about(env!("CARGO_PKG_DESCRIPTION"))
+        .subcommand(SubCommand::with_name("sync").about(
+            "install / update my settings on this computer",
+        ))
         .get_matches();
 
     if let Some(_matches) = matches.subcommand_matches("sync") {